jwt 로그인 및 회원가입  구현

+++ back_gf/src/main/java/com/sol/gf/domain/roles/RolesRepository.java
@@ -7,5 +7,5 @@ import java.util.Optional;
 
 @Repository
 public interface RolesRepository extends JpaRepository<RolesEntity, Long> {
-    Optional<RolesEntity> findByRolesNo(int rolesNo);
+    Optional<RolesEntity> findByRolesNo(long rolesNo);
 }
